==> py/1_dictionary.py <==
#Descriptive Dictionary

#Territorial
TARGET_TERMS = ["district", "back home", "home state", "county", "southern border", "statehouse", "state legislature", "folks back home"]

#Socio_cultural
TARGET_TERMS = ["workers"]

#Occupation
TARGET_TERMS = ["veteran","law enforcement", "sheriff", "doctor", "dentist", "principal", "researcher", "attorney", "police", "mom"]

#Demographics
TARGET_TERMS = ["people","white","women", "woman", "men", "parents", "families","students", "children", "kids", "Black", "Asian", "tribal","christians"]

#Constituencies
TARGET_TERMS = ["small businesses", "border patrol", "veterans","taxpayers"]

# Substansive Dictionary
#Disruptive
TARGET_TERMS = ["object","defund", "dismantle","weaponized", "fraud", "stolen", "deep state", "unelected bureaucrats", "censorship", "subpoena", "protect", "refuse"]

#Maximalist
TARGET_TERMS = ["zero","end"]

#Policy Issue
TARGET_TERMS = ["seurity", "border security", "illegal", "alien", "aliens", "immigration", "migrants", "rape", "crime", "insurance", "inflation","fentanyl", "cartels", "cartel", "trafficking", "crisis", "gas", "taxes", "budget", "debt", "spending", "drugs", "overdoses", "schools", "school", "rights", "gender", "ideology", "health care", "Medicare", "Social Security", "censorship", "401K", "China","life","abortion","DHS"]

#Advocacy
TARGET_TERMS = ["introduce","bill", "demand", "press", "push", "request", "accountable", "subpoena", "oversight", "defend", "protect", "block", "stop"]

#Conventional Cooperation
TARGET_TERMS = ["bipartisan","enacted"]

#General:
TARGET_TERMS = ["law","laws"]

#Own States:
TARGET_TERMS = ["georgia","colorado","arizona","florida","south carolina"]

#Red States:
TARGET_TERMS = ["Alabama", "Alaska", "Arkansas", "Florida", "Idaho", "Indiana", "Iowa", "Kansas", "Kentucky", "Louisiana", "Mississippi", "Missouri", "Montana", "Nebraska", "North Dakota", "Oklahoma", "South Carolina", "South Dakota", "Tennessee", "Texas", "Utah", "West Virginia", "Wyoming"]

#Blue States:
TARGET_TERMS = ["California", "Colorado", "Connecticut", "Delaware", "Hawaii", "Illinois", "Maine", "Maryland", "Massachusetts", "New Jersey", "New Mexico", "New York", "Oregon", "Rhode Island", "Vermont", "Washington"]

#Contested States
TARGET_TERMS = ["Arizona", "Georgia", "Michigan", "Minnesota", "Nevada", "New Hampshire", "North Carolina", "Ohio", "Pennsylvania", "Virginia", "Wisconsin"]
